requirements.txt
# TESTRADE Telemetry Hub Requirements
# Python 3.11+ recommended

# Core dependencies
pyzmq==25.1.2          # ZeroMQ for inter-process communication
redis==5.0.1           # Redis client for data persistence
psutil==5.9.8          # System monitoring and process management

# Web/API dependencies (for GUI backend)
fastapi==0.109.0       # Web framework for GUI backend
uvicorn==0.27.0        # ASGI server for FastAPI
websockets==12.0       # WebSocket support for real-time updates
aiofiles==23.2.1       # Async file operations
aioredis==2.0.1        # Async Redis client for GUI

# Data processing
numpy==1.26.3          # Numerical operations
pandas==2.1.4          # Data manipulation (for GUI analytics)

# Utilities
python-dotenv==1.0.0   # Environment variable management
pydantic==2.5.3        # Data validation (FastAPI dependency)

# Monitoring & Logging
structlog==24.1.0      # Structured logging (optional, but recommended)

# Development/Testing (optional)
pytest==7.4.4
pytest-asyncio==0.21.1
black==23.12.1
mypy==1.8.0